+++ src/main/java/net/berserker_rpg/config/EffectsConfig.java
@@ -6,7 +6,7 @@ public class EffectsConfig {
     public float rage_attack_speed_increase_per_stack = 0.025f;
     public int rage_max_amplifier_stack = 10;
     public float soul_devourer_heal_on_kill_target_max_health_amount = 0.10F;
-    public float bloody_strike_self_damage = 0.8F;
-    public float blood_reckoning_absoprtion_to_heal = 0.6F;
+    public float bloody_strike_self_damage = 1.0F;
+    public float blood_reckoning_absoprtion_to_heal = 0.4F;
     public float blood_reckoning_base_heal = 4.0F;
 }